sincronizando branches

+++ IntegracaoInterfaces/setDados/setDatasValuation.py
@@ -5,7 +5,7 @@ def extValuation(dados):
     ebit = [float(v) for (k ,v) in dados.items() if re.search("\\bebit.\\b", k)]
     ebitda = [float(v) for (k ,v) in  dados.items() if "ebitda" in k ]
     ncg = [float(v) for (k ,v) in dados.items() if "cg" in k]
-    stocks = int(dados['stocks'])
+    stocks = float(dados['stocks'])
     ir = int(dados['IR'])
     growth = int(dados['growth'])
     reinvest = float(dados['reinvest'])
